+++ src/pages/Write.tsx
@@ -9,10 +9,10 @@ import { ReactComponent as Back } from '../assets/svg/back_24_B.svg';
 import { ReactComponent as Check } from '../assets/svg/check_24.svg';
 import { ReactComponent as UnCheck } from '../assets/svg/unCheck_24.svg';
 import { ReactComponent as Xmark } from '../assets/svg/x.svg';
+import Divider from '../components/@common/Divider/Divider';
 import Text from '../components/@common/Text/Text';
 import BottomSheet from '../components/BottomSheet/BottomSheet';
 import { checkboxOptions, getDomain } from '../components/BottomSheet/CheckBox';
-import Divider from '../components/Divider';
 import Checkbox from '../components/Inputs/Checkbox';
 import Dropdown from '../components/Inputs/Dropdown/Dropdown';
 import Radio, { radioOptions } from '../components/Inputs/Radio';
